+++ frontend/src/api/trips/trips.ts
@@ -1,5 +1,5 @@
 import { END_POINTS } from '@constants/api';
-import type { TripsData } from '@type/trips';
+import type { NewTripData, TripsData } from '@type/trips';
 
 import { axiosInstance } from '@api/axiosInstance';
 
@@ -7,4 +7,10 @@ export const getTrips = async () => {
   const { data } = await axiosInstance.get<{ trips: TripsData[] }>(END_POINTS.TRIPS);
 
   return data?.trips;
+};
+
+export const postTrips = async (newTripData: NewTripData) => async () => {
+  const response = await axiosInstance.post(END_POINTS.TRIPS, newTripData);
+
+  return response;
 };
