+++ frontend/src/components/newTrip/NewTripForm/NewTripForm.style.ts
@@ -1,0 +1,10 @@
+import { css } from '@emotion/react';
+import { Theme } from 'hang-log-design-system';
+
+export const formStyling = css({
+  width: '100%',
+  display: 'flex',
+  flexDirection: 'column',
+  gap: Theme.spacer.spacing4,
+  marginTop: Theme.spacer.spacing5,
+});
