+++ frontend/src/hooks/newTrip/useNewTripForm.ts
@@ -1,0 +1,38 @@
+import { CityData } from '@type/city';
+import { DateRangeData, NewTripData } from '@type/trips';
+import { useEffect, useState } from 'react';
+
+const initialNewTripData = {
+  startDate: null,
+  endDate: null,
+  cityIds: [],
+};
+
+export const useNewTripForm = () => {
+  const [newTripData, setNewTripData] = useState<NewTripData>(initialNewTripData);
+  const [isAllInputFilled, setIsAllInputFilled] = useState(false);
+
+  useEffect(() => {
+    validateInputs();
+  }, [newTripData]);
+
+  const setCityData = (cities: CityData[]) => {
+    const cityIds = cities.map((city) => city.id);
+
+    setNewTripData((prev) => ({ ...prev, cityIds }));
+  };
+
+  const setDateData = (dateRange: DateRangeData) => {
+    const { start: startDate, end: endDate } = dateRange;
+
+    setNewTripData((prev) => ({ ...prev, startDate, endDate }));
+  };
+
+  const validateInputs = () => {
+    const { startDate, endDate, cityIds } = newTripData;
+
+    setIsAllInputFilled(!!startDate && !!endDate && !!cityIds.length);
+  };
+
+  return { newTripData, setCityData, setDateData, isAllInputFilled };
+};
